+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Treat scheduled outgoing activities where sender is not specified as invalid.
 - Removed unhashed OAuth tokens from database.
 - Don't change default config file path when `production` feature is enabled.
+- Changed default data retention parameters.
 
 ### Removed